# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Library AI — Book recommendation API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SimpleStatus {
  ok: boolean;
}

export type UserReadingStyle =
  | (typeof UserReadingStyle)[keyof typeof UserReadingStyle]
  | null;

export const UserReadingStyle = {
  fast_paced: "fast_paced",
  academic: "academic",
  casual: "casual",
} as const;

export interface User {
  id: string;
  name: string;
  email: string;
  phone?: string | null;
  favoriteGenres: string[];
  favoriteAuthors: string[];
  readingStyle?: UserReadingStyle;
  onboarded: boolean;
}

export interface CurrentUser {
  user: User | null;
}

export interface SignupRequest {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  phone?: string | null;
  favoriteGenres?: string[];
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UpdatePreferencesRequestReadingStyle =
  (typeof UpdatePreferencesRequestReadingStyle)[keyof typeof UpdatePreferencesRequestReadingStyle];

export const UpdatePreferencesRequestReadingStyle = {
  fast_paced: "fast_paced",
  academic: "academic",
  casual: "casual",
} as const;

export interface UpdatePreferencesRequest {
  favoriteGenres: string[];
  favoriteAuthors: string[];
  readingStyle: UpdatePreferencesRequestReadingStyle;
}

export interface Book {
  id: string;
  title: string;
  author: string;
  coverUrl: string;
  genre: string;
  tags: string[];
  rating: number;
  ratingsCount: number;
  pages: number;
  year: number;
  shortDescription: string;
}

export type BookDetailLibraryStatus =
  | (typeof BookDetailLibraryStatus)[keyof typeof BookDetailLibraryStatus]
  | null;

export const BookDetailLibraryStatus = {
  reading: "reading",
  read: "read",
  want_to_read: "want_to_read",
} as const;

export type BookDetail = Book & {
  description: string;
  authorBio: string;
  similarBooks: Book[];
  inLibrary: boolean;
  libraryStatus?: BookDetailLibraryStatus;
  userRating?: number | null;
};

export type LibraryEntryStatus =
  (typeof LibraryEntryStatus)[keyof typeof LibraryEntryStatus];

export const LibraryEntryStatus = {
  reading: "reading",
  read: "read",
  want_to_read: "want_to_read",
} as const;

export interface LibraryEntry {
  id: string;
  bookId: string;
  book: Book;
  status: LibraryEntryStatus;
  /** 0-100 percent */
  progress: number;
  rating?: number | null;
  addedAt: string;
}

export type AddLibraryEntryRequestStatus =
  (typeof AddLibraryEntryRequestStatus)[keyof typeof AddLibraryEntryRequestStatus];

export const AddLibraryEntryRequestStatus = {
  reading: "reading",
  read: "read",
  want_to_read: "want_to_read",
} as const;

export interface AddLibraryEntryRequest {
  bookId: string;
  status: AddLibraryEntryRequestStatus;
}

export type UpdateLibraryEntryRequestStatus =
  (typeof UpdateLibraryEntryRequestStatus)[keyof typeof UpdateLibraryEntryRequestStatus];

export const UpdateLibraryEntryRequestStatus = {
  reading: "reading",
  read: "read",
  want_to_read: "want_to_read",
} as const;

export interface UpdateLibraryEntryRequest {
  status?: UpdateLibraryEntryRequestStatus;
  progress?: number;
  rating?: number;
}

export type RecommendationModel =
  (typeof RecommendationModel)[keyof typeof RecommendationModel];

export const RecommendationModel = {
  content: "content",
  collaborative: "collaborative",
  hybrid: "hybrid",
} as const;

export interface Recommendation {
  book: Book;
  /** 0-100 AI match score */
  matchPercent: number;
  /** Human-readable explanation */
  reason: string;
  model: RecommendationModel;
}

export type ProfileInsightsGenreBreakdownItem = {
  genre: string;
  count: number;
};

export interface ProfileInsights {
  booksRead: number;
  currentlyReading: number;
  wantToRead: number;
  readingTimeHours: number;
  favoriteGenre: string;
  averageRating: number;
  aiInsight: string;
  genreBreakdown: ProfileInsightsGenreBreakdownItem[];
}

export type ListBooksParams = {
  genre?: string;
  author?: string;
  minRating?: number;
  search?: string;
  limit?: number;
};

export type ListLibraryParams = {
  status?: ListLibraryStatus;
};

export type ListLibraryStatus =
  (typeof ListLibraryStatus)[keyof typeof ListLibraryStatus];

export const ListLibraryStatus = {
  reading: "reading",
  read: "read",
  want_to_read: "want_to_read",
} as const;

export type GetRecommendationsParams = {
  model?: GetRecommendationsModel;
  limit?: number;
};

export type GetRecommendationsModel =
  (typeof GetRecommendationsModel)[keyof typeof GetRecommendationsModel];

export const GetRecommendationsModel = {
  content: "content",
  collaborative: "collaborative",
  hybrid: "hybrid",
} as const;