# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Library AI — Book recommendation API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication and current user
  - name: users
    description: User profile and onboarding
  - name: books
    description: Book catalog
  - name: library
    description: User library entries
  - name: recommendations
    description: AI recommendations
  - name: insights
    description: Reading insights and stats
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Create an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in to an existing account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Sign out the current session
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleStatus"

  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the currently authenticated user
      responses:
        "200":
          description: Current user (or null)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"

  /users/preferences:
    put:
      operationId: updatePreferences
      tags: [users]
      summary: Save onboarding preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePreferencesRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /books:
    get:
      operationId: listBooks
      tags: [books]
      summary: Browse the book catalog
      parameters:
        - in: query
          name: genre
          schema:
            type: string
        - in: query
          name: author
          schema:
            type: string
        - in: query
          name: minRating
          schema:
            type: number
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Book"

  /books/trending:
    get:
      operationId: listTrendingBooks
      tags: [books]
      summary: Trending books across the platform
      responses:
        "200":
          description: Trending books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Book"

  /books/{bookId}:
    get:
      operationId: getBook
      tags: [books]
      summary: Get a book by id
      parameters:
        - in: path
          name: bookId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Book details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookDetail"

  /library:
    get:
      operationId: listLibrary
      tags: [library]
      summary: List the current user's library
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [reading, read, want_to_read]
      responses:
        "200":
          description: Library entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LibraryEntry"

  /library/continue-reading:
    get:
      operationId: listContinueReading
      tags: [library]
      summary: Books the user is currently reading
      responses:
        "200":
          description: In-progress books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LibraryEntry"

  /library/entries:
    post:
      operationId: addToLibrary
      tags: [library]
      summary: Add a book to the user's library
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddLibraryEntryRequest"
      responses:
        "200":
          description: Created entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryEntry"

  /library/entries/{bookId}:
    patch:
      operationId: updateLibraryEntry
      tags: [library]
      summary: Update status / rating / progress
      parameters:
        - in: path
          name: bookId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLibraryEntryRequest"
      responses:
        "200":
          description: Updated entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryEntry"
    delete:
      operationId: removeFromLibrary
      tags: [library]
      summary: Remove a book from the user's library
      parameters:
        - in: path
          name: bookId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleStatus"

  /recommendations:
    get:
      operationId: getRecommendations
      tags: [recommendations]
      summary: AI recommendations for the user
      parameters:
        - in: query
          name: model
          schema:
            type: string
            enum: [content, collaborative, hybrid]
            default: hybrid
        - in: query
          name: limit
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: Recommended books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recommendation"

  /recommendations/based-on-interests:
    get:
      operationId: getBasedOnInterests
      tags: [recommendations]
      summary: Recommendations based on saved genre interests
      responses:
        "200":
          description: Books matching interests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recommendation"

  /insights/profile:
    get:
      operationId: getProfileInsights
      tags: [insights]
      summary: AI-generated profile insights and reading stats
      responses:
        "200":
          description: Insights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileInsights"

  /insights/genres:
    get:
      operationId: getGenreList
      tags: [insights]
      summary: List of all genres in the catalog
      responses:
        "200":
          description: Genres
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    SimpleStatus:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]

    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        favoriteGenres:
          type: array
          items:
            type: string
        favoriteAuthors:
          type: array
          items:
            type: string
        readingStyle:
          type: string
          nullable: true
          enum: [fast_paced, academic, casual, null]
        onboarded:
          type: boolean
      required: [id, name, email, favoriteGenres, favoriteAuthors, onboarded]

    CurrentUser:
      type: object
      properties:
        user:
          oneOf:
            - $ref: "#/components/schemas/User"
            - type: "null"
      required: [user]

    SignupRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        phone:
          type: string
          nullable: true
        favoriteGenres:
          type: array
          items:
            type: string
      required: [name, email, password]

    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]

    UpdatePreferencesRequest:
      type: object
      properties:
        favoriteGenres:
          type: array
          items:
            type: string
        favoriteAuthors:
          type: array
          items:
            type: string
        readingStyle:
          type: string
          enum: [fast_paced, academic, casual]
      required: [favoriteGenres, favoriteAuthors, readingStyle]

    Book:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        author:
          type: string
        coverUrl:
          type: string
        genre:
          type: string
        tags:
          type: array
          items:
            type: string
        rating:
          type: number
        ratingsCount:
          type: integer
        pages:
          type: integer
        year:
          type: integer
        shortDescription:
          type: string
      required: [id, title, author, coverUrl, genre, tags, rating, ratingsCount, pages, year, shortDescription]

    BookDetail:
      allOf:
        - $ref: "#/components/schemas/Book"
        - type: object
          properties:
            description:
              type: string
            authorBio:
              type: string
            similarBooks:
              type: array
              items:
                $ref: "#/components/schemas/Book"
            inLibrary:
              type: boolean
            libraryStatus:
              type: string
              nullable: true
              enum: [reading, read, want_to_read, null]
            userRating:
              type: number
              nullable: true
          required: [description, authorBio, similarBooks, inLibrary]

    LibraryEntry:
      type: object
      properties:
        id:
          type: string
        bookId:
          type: string
        book:
          $ref: "#/components/schemas/Book"
        status:
          type: string
          enum: [reading, read, want_to_read]
        progress:
          type: integer
          description: 0-100 percent
        rating:
          type: number
          nullable: true
        addedAt:
          type: string
          format: date-time
      required: [id, bookId, book, status, progress, addedAt]

    AddLibraryEntryRequest:
      type: object
      properties:
        bookId:
          type: string
        status:
          type: string
          enum: [reading, read, want_to_read]
      required: [bookId, status]

    UpdateLibraryEntryRequest:
      type: object
      properties:
        status:
          type: string
          enum: [reading, read, want_to_read]
        progress:
          type: integer
        rating:
          type: number

    Recommendation:
      type: object
      properties:
        book:
          $ref: "#/components/schemas/Book"
        matchPercent:
          type: integer
          description: 0-100 AI match score
        reason:
          type: string
          description: Human-readable explanation
        model:
          type: string
          enum: [content, collaborative, hybrid]
      required: [book, matchPercent, reason, model]

    ProfileInsights:
      type: object
      properties:
        booksRead:
          type: integer
        currentlyReading:
          type: integer
        wantToRead:
          type: integer
        readingTimeHours:
          type: integer
        favoriteGenre:
          type: string
        averageRating:
          type: number
        aiInsight:
          type: string
        genreBreakdown:
          type: array
          items:
            type: object
            properties:
              genre:
                type: string
              count:
                type: integer
            required: [genre, count]
      required: [booksRead, currentlyReading, wantToRead, readingTimeHours, favoriteGenre, averageRating, aiInsight, genreBreakdown]