# lib/api-zod/src/generated/types/bookDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Library AI — Book recommendation API
 * OpenAPI spec version: 0.1.0
 */
import type { Book } from "./book";
import type { BookDetailLibraryStatus } from "./bookDetailLibraryStatus";

export type BookDetail = Book & {
  description: string;
  authorBio: string;
  similarBooks: Book[];
  inLibrary: boolean;
  libraryStatus?: BookDetailLibraryStatus;
  userRating?: number | null;
};