# lib/api-zod/src/generated/types/profileInsights.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Library AI — Book recommendation API
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileInsightsGenreBreakdownItem } from "./profileInsightsGenreBreakdownItem";

export interface ProfileInsights {
  booksRead: number;
  currentlyReading: number;
  wantToRead: number;
  readingTimeHours: number;
  favoriteGenre: string;
  averageRating: number;
  aiInsight: string;
  genreBreakdown: ProfileInsightsGenreBreakdownItem[];
}