# lib/api-zod/src/generated/types/recommendation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Library AI — Book recommendation API
 * OpenAPI spec version: 0.1.0
 */
import type { Book } from "./book";
import type { RecommendationModel } from "./recommendationModel";

export interface Recommendation {
  book: Book;
  /** 0-100 AI match score */
  matchPercent: number;
  /** Human-readable explanation */
  reason: string;
  model: RecommendationModel;
}