# lib/api-zod/src/generated/types/updateLibraryEntryRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Library AI — Book recommendation API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateLibraryEntryRequestStatus } from "./updateLibraryEntryRequestStatus";

export interface UpdateLibraryEntryRequest {
  status?: UpdateLibraryEntryRequestStatus;
  progress?: number;
  rating?: number;
}